+++ domain/buildAITerminalSessionInfo.test.ts
@@ -1,7 +1,7 @@
 import test from "node:test";
 import assert from "node:assert/strict";
 
-import { buildAITerminalSessionInfo } from "./buildAITerminalSessionInfo";
+import { buildAITerminalSessionInfo } from "./buildAITerminalSessionInfo.ts";
 import type { Host, TerminalSession } from "../../types";
 
 const baseHost = (overrides: Partial<Host> = {}): Host =>
